+++ FarNet/FarNetMan/UI.h
@@ -31,7 +31,7 @@ public:
 	virtual ConsoleColor GetPaletteBackground(PaletteColor paletteColor) override;
 	virtual ConsoleColor GetPaletteForeground(PaletteColor paletteColor) override;
 	virtual int ReadKeys(array<KeyData^>^ keys) override;
-	virtual int SaveScreen(int x1, int y1, int x2, int y2) override;
+	virtual IntPtr SaveScreen(int x1, int y1, int x2, int y2) override;
 	virtual KeyInfo^ ReadKey(ReadKeyOptions options) override;
 	virtual String^ GetBufferLineText(int lineIndex) override;
 	virtual void Break() override;
@@ -41,7 +41,7 @@ public:
 	virtual void DrawPalette(int left, int top, PaletteColor paletteColor, String^ text) override;
 	virtual void FlushInputBuffer() override;
 	virtual void Redraw() override;
-	virtual void RestoreScreen(int screen) override;
+	virtual void RestoreScreen(IntPtr screen) override;
 	virtual void SaveUserScreen() override;
 	virtual void SetProgressFlash() override;
 	virtual void SetProgressState(TaskbarProgressBarState state) override;
